+++ common/src/main/java/net/mca/entity/ai/relationship/Gender.java
@@ -52,34 +52,12 @@ public enum Gender {
         return dataName;
     }
 
-    public boolean isNonBinary() {
-        return this == NEUTRAL || this == UNASSIGNED;
-    }
-
-    public Stream<Gender> getTransients() {
-        return isNonBinary() ? Stream.of(MALE, FEMALE) : Stream.of(this);
-    }
-
     public Gender binary() {
         return this == FEMALE ? FEMALE : MALE;
     }
 
     public Gender opposite() {
         return this == FEMALE ? MALE : FEMALE;
-    }
-
-    /**
-     * Checks whether this gender is attracted to another.
-     */
-    public boolean isAttractedTo(Gender other) {
-        return other == UNASSIGNED || this == NEUTRAL || other != this;
-    }
-
-    /**
-     * Checks whether both genders are mutually attracted to each other.
-     */
-    public boolean isMutuallyAttracted(Gender other) {
-        return isAttractedTo(other) && other.isAttractedTo(this);
     }
 
     public static Gender byId(int id) {